requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.6
django-heroku==0.3.1
django-ip==1.0.2
django-rest-framework==0.1.0
django-six==1.0.4
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
gunicorn==19.9.0
idna==2.8
psycopg2-binary==2.8.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.2
whitenoise==4.1.2
